+++ src/schema/resolvers/CartItem/Mutation/index.ts
@@ -1,0 +1,5 @@
+import updateQuantity from './updateQuantity'
+
+export default {
+  updateQuantity,
+}
